requirements.txt
aniso8601==10.0.1
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
et_xmlfile==2.0.0
Flask==2.3.3
Flask-Login==0.6.3
flask-restx==1.3.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==2.3.7
WTForms==3.1.1
xlrd==2.0.2
